+++ addons/lux/editor/lux_editor_plugin.gd
@@ -12,6 +12,9 @@ var _dock: Control
 
 func _enter_tree() -> void:
 	add_custom_type("LuxRoot", "Node3D", preload("res://addons/lux/runtime/lux_root.gd"), RootIcon)
+	add_custom_type(
+		"LuxRoleTag", "Node", preload("res://addons/lux/runtime/lux_role_tag.gd"), RootIcon
+	)
 	_dock = DockScene.instantiate()
 	_dock.set_editor_interface(get_editor_interface())
 	add_control_to_dock(EditorPlugin.DOCK_SLOT_RIGHT_UL, _dock)
@@ -22,4 +25,5 @@ func _exit_tree() -> void:
 		remove_control_from_docks(_dock)
 		_dock.free()
 		_dock = null
+	remove_custom_type("LuxRoleTag")
 	remove_custom_type("LuxRoot")